Move MaterialPropertyInstance into Gameplay.Visuals
