+++ Assets/Script/Gameplay/Visuals/MaterialPropertyInstance.cs.meta
