--- Assets/Script/Rendering.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5450a5a5c17360152a2d27ac013328ba
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
